client: edit structure (replace some components with views)
