+++ client/src/router/router.ts
@@ -1,6 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 import LoginView from '../views/LoginView.vue'
+import ResetView from '../views/ResetView.vue'
 import RegisterView from '../views/RegisterView.vue'
 import ProfileView from '../views/ProfileView.vue'
 import HistoryView from '../views/HistoryView.vue'
@@ -19,6 +20,11 @@ const router = createRouter({
       path: '/login/:confirm?',
       name: 'login',
       component: LoginView
+    },
+    {
+      path: '/reset/:challenge?',
+      name: 'reset',
+      component: ResetView
     },
     {
       path: '/register',
