+++ client/src/views/ResetView.vue
@@ -1,0 +1,13 @@
+<template>
+
+</template>
+
+<script lang="ts">
+export default {
+  name: "ResetView"
+};
+</script>
+
+<style scoped>
+
+</style>
